Delete Modified
Delete now has an error checking for deleting existing items currently in transaction.

+++ delete_chemical_item.php
@@ -14,7 +14,23 @@
 	$MyConnection = mysqli_connect($MyServer, $MyUserName, $MyPassword, $MyDBName);
 
 	$item = $_POST['CHEM_ID'];
-	
 
-	mysqli_query($MyConnection, "DELETE FROM chemicals WHERE (chemicals.Chemical_Id = $item)");
+	$MySearchQuery = "SELECT * FROM transaction WHERE (Chemical_Id = $item);";
+	$MyValues = mysqli_query($MyConnection, $MySearchQuery);
+
+	if (mysqli_num_rows($MyValues) > 0)
+	{
+		echo json_encode(array(
+    		'status' => 'error'
+		));
+	}
+
+	else
+	{
+		mysqli_query($MyConnection, "DELETE FROM chemicals WHERE (chemicals.Chemical_Id = $item)");
+		echo json_encode(array(
+    		'status' => 'success'
+		));
+	}
+	
 ?>
